Tell rm not to prompt
When recompress is combined with tar_scm with scm set to "git" and package-meta
set to "yes", the tarball produced by tar_scm contains the .git directory,
which in turn contains lots of write-protected files. Not supplying the -f
option causes rm to ask - for each file - if it really should remove it.

Signed-off-by: Nathan Cutler <ncutler@suse.cz>