Change default compress level
Set the default compress level from 15 to 19. On packages that are quite
large, especially common in rust this makes a large difference is the
size of the resulting source.

For example in Kanidm the difference between 15 to 19 is 69MB to 60MB. Due
to OBS' limited bandwidth, it can be extremely slow to download these sources
from OBS, so that 9MB saving results in downloads being 90s faster to complete.